// types.ts
export interface Service {
  icon: string;
  title: string;
  description: string;
}

export interface PortfolioItem {
  image: string;
  title: string;
  description: string;
  client: string;
  metric: string;
}

export interface Testimonial {
  quote: string;
  author: string;
  company: string;
  image: string;
}